pyproject.toml
[tool]
[tool.poetry]
name = "igit"
version = "0.1.0"
homepage = "https://github.com/jmosbacher/igit"
description = "Top-level package for igit."
authors = ["Yossi Mosbacher <joe.mosbacher@gmail.com>"]
readme = "README.rst"
license = "Apache-2.0"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
]
packages = [
    { include = "igit" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.7"
click = "*"
intervaltree = "^3.1.0"
treelib = "^1.6.1"
pydantic = "^1.8.1"
fsspec = "^2021.4.0"
zict = "^2.0.0"
aiohttp = {version = "^3.7.4", extras = ["server"]}
requests = {version = "^2.25.1", extras = ["server"]}
Jinja2 = {version = "^2.11.3", extras = ["server"]}
fastapi = {version = "^0.63.0", extras = ["server"]}
msgpack = "^1.0.2"
hypothesis = {extras = ["numpy", "pandas"], version = "^6.10.1"}
cryptography = "^3.4.7"
pymongo = "^3.11.3"
dill = "^0.3.3"
msgpack-numpy = "^0.4.7"

[tool.poetry.dev-dependencies]
bumpversion = "*"
coverage = "*"
flake8 = "*"
isort = "*"
pylint = "*"
dephell = "*"
pytest = "*"
sphinx = "*"
tox = "*"
yapf = "*"
sphinx-material = "*"
nbsphinx = "*"
invoke = "*"

[tool.poetry.scripts]
igit = 'igit.cli:main'

[build-system]
requires = ["poetry>=0.12", "setuptools"]
build-backend = "poetry.masonry.api"

[tool.dephell.main]
versioning = "semver"
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}
